Add timestamp in nanoseconds for influxdb output
For the "Closed yesterday" query we can tell telegraf the exact timestamp
for the data.
This way we can execute the telegraf query more often per day to ensure
we get data, and also for the other queries there's no need to run this
only once a day.

Ticket: https://progress.opensuse.org/issues/121582#note-57