Include "Feedback" in cycle time as well
Only counting "In Progress" yields unrealistically low cycle time values
and would not reflect overall time that a common development ticket is
being worked on within a team as often we use "Feedback" when waiting
for others within the team, e.g.  for code review.

This commit counts time in status "Feedback" same as "In Progress" to
account for that.

The test adaptions include the definition of "Feedback" in the mocked
data as well as a mock entry for a ticket in "Feedback" with according
adaption in the assertion but no assertion is added for the calculation
to include tickets which had transitioned over that state.

Related progress issue: https://progress.opensuse.org/issues/155917