Run preview.yaml without secret key
To avoid the use of secret variable in CI which breaks the pr-preview-action
when it runs from forked repo, the preview.yaml is updated to use
pull_request_target event instead of pull_request. Permissions are removed as
the GITHUB_TOKEN is granted read/write repository permission by default for
that event.

https://progress.opensuse.org/issues/158236

Signed-off-by: ybonatakis <ybonatakis@suse.com>